More algo fixes...
But the mysterious illegal memory access remains

+++ src/crypto/hash/cuda/config.h
@@ -9,7 +9,7 @@
 
 typedef unsigned char BYTE;
 typedef unsigned int  WORD;
-typedef uint64_t LONG;
+typedef unsigned long long LONG;
 
 #include <stdlib.h>
 #include <string.h>
